move CorsConfig to separate module
CorsConfig class is still available in aiohttp_cors module
